complete homeview
