+++ Fitness App/Home/Models/Workout.swift
@@ -1,10 +1,11 @@
 //
-//  WorkoutCard.swift
+//  Workout.swift
 //  Fitness App
 //
 //  Created by Admin on 05/09/2024.
 //
 
+import Foundation
 import SwiftUI
 
 
@@ -25,38 +26,4 @@ struct Workout{
                         Workout(id: 3, title: "Running", image: "figure.run", tintColor: .gray, duration: "23 mins", date: "October 12, 2024", calories: "21 Kcal"),
                         Workout(id: 4, title: "Running", image: "figure.run", tintColor: .red, duration: "48 mins", date: "Aprial 3, 2024", calories: "23 Kcal")]
     
-}
-
-struct WorkoutCard: View {
-    
-    @State var workout : Workout
-    var body: some View {
-
-        HStack{
-            Image(systemName: workout.image).resizable().scaledToFit()
-                .frame(width: 40, height: 40)
-                .foregroundStyle(workout.tintColor)
-                .padding()
-                .background(.gray.opacity(0.1))
-                .cornerRadius(10)
-            
-            VStack(spacing : 16){
-                HStack{
-                    Text(workout.title).font(.title3).bold()
-                    Spacer()
-                    Text(workout.duration)
-                }
-                
-                HStack{
-                    Text(workout.date)
-                    Spacer()
-                    Text(workout.calories)
-                }
-            }
-        }.padding(.horizontal)
-    }
-}
-
-#Preview {
-    WorkoutCard(workout: .example)
 }
